code cleanup: remove unnecessary casts, simplify conditions

+++ src/pal/datatype/NumericDataType.java
@@ -80,7 +80,7 @@ public class NumericDataType extends SimpleDataType {
      * Returns numeric index (size) from a unique ascii character
      */
     public final int getNumericIndexFromNumericChar(char c) {
-        return (int) (c - 64);
+        return (c - 64);
     }
 
     /**
